% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_stan_code.R
\name{print_stan_code}
\alias{print_stan_code}
\title{prints a section of stan code to console}
\usage{
print_stan_code(stan_code, section = NULL)
}
\arguments{
\item{stan_code}{(chr) text of stan code}

\item{section}{(chr) name of section to print. Should be one of 'data', 'parameters', 'functions', etc}
}
\description{
prints a section of stan code to console
}

