% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_stan_code.R
\name{print_stan_file}
\alias{print_stan_file}
\title{prints stan file to the console}
\usage{
print_stan_file(stan_file, ...)
}
\arguments{
\item{stan_file}{path to stan file}

\item{...}{(params to print_stan_code - ex: section)}
}
\value{
character scalar containing stan code
}
\description{
prints stan file to the console
}

