% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_stan_code.R
\name{read_stan_file}
\alias{read_stan_file}
\title{reads a stan file into R}
\usage{
read_stan_file(stan_file)
}
\arguments{
\item{path}{to stan file}
}
\value{
character scalar containing stan code
}
\description{
reads a stan file into R
}

